Develop -> Stable

+++ src/Regula.DocumentReader.WebClient/Model/RfidSecurityObject.cs
@@ -45,7 +45,7 @@ namespace Regula.DocumentReader.WebClient.Model
         /// <param name="fileReference">Reference to the source file of the security object data (required).</param>
         /// <param name="notifications">List of remarks arisen during the analysis of SO data structure. Can be ParsingErrorCodes or ParsingNotificationCodes enum. (required).</param>
         /// <param name="signerInfos">List of containers to store information about digital signature objects contained in the SO (required).</param>
-        public RfidSecurityObject(decimal varVersion = default(decimal), string objectType = default(string), decimal fileReference = default(decimal), List<int> notifications = default(List<int>), List<RfidSignerInfoEx> signerInfos = default(List<RfidSignerInfoEx>))
+        public RfidSecurityObject(decimal varVersion = default(decimal), string objectType = default(string), decimal fileReference = default(decimal), List<long> notifications = default(List<long>), List<RfidSignerInfoEx> signerInfos = default(List<RfidSignerInfoEx>))
         {
             this.VarVersion = varVersion;
             // to ensure "objectType" is required (not null)
@@ -95,7 +95,7 @@ namespace Regula.DocumentReader.WebClient.Model
         /// </summary>
         /// <value>List of remarks arisen during the analysis of SO data structure. Can be ParsingErrorCodes or ParsingNotificationCodes enum.</value>
         [DataMember(Name = "Notifications", IsRequired = true, EmitDefaultValue = true)]
-        public List<int> Notifications { get; set; }
+        public List<long> Notifications { get; set; }
 
         /// <summary>
         /// List of containers to store information about digital signature objects contained in the SO
